ewaol-tests: Include STDERR in the logs for failing OCI runtime tests
To avoid contaminating the STDOUT stream returned from a Bats run function, we
redirect the STDERR of each OCI container command. Rather than redirecting
STDERR to /dev/null, this patch stores it in a temporary log file within the
OCI_TESTS_LOG_DIR directory.

If there is a test failure, this file's contents are then included in the
debugging output so that the error can be more easily investigated.

This patch also simplifies the logging to avoid significant code duplication,
and generally improves the scripts following shellcheck static analysis.

Issue-Id: SCM-2619
Signed-off-by: Richard Neill <richard.neill@arm.com>
Signed-off-by: Diego Sueiro <diego.sueiro@arm.com>
Change-Id: I8be08b44e58d8f718e4b915296bd55785d840bd4