ewaol-tests: Add meta-ewaol-tests layer with OCI container integration tests
Adds the meta-ewaol-tests layer to support the validation of images built for
the EWAOL project.

This patch adds Bats-based runtime integration testing for OCI containers. The
integration tests are provided via the oci-runtime-integration-tests.bb recipe.
This recipe deploys tests to validate the ability for a container to execute an
example persistent workload, which involves the pulling of an image from an
image hub on the network, the creation and start of a detached container, and
its subsequent stop and removal.

The tests are made available to run with a standalone shell script via the
run-oci-runtime-integration-tests binary, and as a suite integrated with the
Yocto Package Test (ptest) framework using ptest-runner.

The tests can be included in the build by appending 'ewaol-test' to the build
configuration's DISTRO_FEATURES. A build configuration file
meta-ewaol-config/kas/tests.yml is included to support this inclusion via the
kas build tool.

Issue-Id: SCM-2357
Signed-off-by: Richard Neill <richard.neill@arm.com>
Signed-off-by: Diego Sueiro <diego.sueiro@arm.com>
Change-Id: I2659e27fca12ae2cf9b9d0cd13041eafced7ebe1