ewaol-config/ci: Provide YAML file to define kas build targets
Provide a YAML file to define build targets (build machines, build modifiers,
CI builds) that can be queried via the kas build script. The list of CI build
targets will be built when "all" is passed to the build script.

Also provide a configuration file to configure a CI build in which the
workspace is subsequently cleaned to conserve disk usage.

Issue-Id: SCM-2619
Signed-off-by: Richard Neill <richard.neill@arm.com>
Signed-off-by: Diego Sueiro <diego.sueiro@arm.com>
Change-Id: I4e5f70734caf72d7da6133e60cebec67034b76ef