tools: Provide generic class for executing Python scripts within a venv
This patch adds a top-level directory to support tooling within the repository.

To this directory, this patch refactors the virtual environment component of
the code checking tool into a generic ModulesVirtualEnv class that may be used
to generally run Python scripts within a venv.

ModulesVirtualEnv takes a runner script, arguments, and the Python package
dependencies required to run the script with those arguments. A Python virtual
environment is then created, in which the dependencies are installed, and the
script is called from that virtual environment context as a subprocess.

Issue-Id: SCM-2584
Issue-Id: SCM-2585
Change-Id: I4835c5e3e02ccd448de4dcedeee3f3e61f4a8a13
Signed-off-by: Richard Neill <richard.neill@arm.com>