tools/qa-checks: Add runner script for checker modules
This patch adds a runner script that enables execution of a suite of checkers,
provided as command-line arguments. The checkers are modular, and must
implement the functions defined by a helper abstract class given in
abstract_check.py in order to be run by the runner script.

By default, the checkers run within a Python virtual environment using the
ModulesVirtualEnv class.

Usage information can be found by passing --help to run-checks.py.

Issue-Id: SCM-2584
Issue-Id: SCM-2585
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ia2e07ef14e9b01da25893557ef81b197d2f793ff