ewaol-tests: Handle incomplete prior test-suite execution
This patch handles the case where a previous test suite may fail in a
hanging state, be user-interrupted and suspended, or orphaned during its
execution, in order to avoid the possibility of running multiple interleaving
test suites on the system which may interfere and cause the tests to fail
unexpectedly with difficult-to-diagnose errors.

To do this, when a test suite execution begins it writes its PGID to a
run-file, and removes this run-file when it ends. Before any execution
begins, the existance of a run-file is checked in order to determine if a
previous execution is running. If it is running, a SIGINT to sent to the
previous execution's PGID, before continuing.

Issue-Id: SCM-3042
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I44dfeff7c4566e7ce6b0b65c6c71e95238298209