tools/qa-checks: Enable configurable root directory for running the checks
User-provided relative paths are considered relative from the root directory.
This patch allows the user to configure that root directory, so that the checks
may be easily applied to other projects on the file system,

The patch also refactors how keyword values are initialised, making them only
evaluate if they are needed. For example, the contents of the project's
.gitignore will only be read and processed if needed for the check
configuration.

Issue-Id: SCM-3033
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I02a087477942e5b152546b23a01d38f45be004da