tools/qa-checks: Add remaining error handling for missing Python modules
In order to read the configuration file, the host system must have the pyyaml
Python package. As this is needed prior to any virtual environment set up, this
patch notifies the user if it cannot be found and aborts gracefully.

In addition, wrap the module import to warn if email_validator cannot be found
for the commit message QA-check (this should only happen if not using a virtual
environment).

Issue-Id: SCM-3033
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ia564ecdd2a26fa499d1a524232882a32883d4279