tools/qa-checks: Enable the commit message QA-check to analyse patch chains
This patch allows the user to provide a set of commits to the commit message
check, enabling the ability to analyse a patch chain.

To do this, the commit message accepts a 'commits' parameter, which should be
defined in one of the follow two formats:

* '-N' to check the latest N commit messages
* 'commit1(,commit2,...)' as a string list of commits to check. The commits
  must be valid when passed to the 'git show' command, for example a commit SHA
  or a relative commit like HEAD~2.

This patch also excludes lines that are URLs from the message length check, as
URLs are allowed break the maximum character length.

Issue-Id: SCM-3033
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I0a6aa7552206c7e4a963d9ddf0687ff773386957