doc: Do not embed README.md in the Sphinx-rendered documentation
This patch decouples README.md from the rendered documentation, allowing its
non-Markdown text to be swapped to Markdown format.

The README and relevant documentation pages are refactored so that the README
only contains the critical information, and references the main content within
the project documentation.

The Sphinx package 'm2r2' which supports embedding .md files within .rst files
is no longer necessary, and is removed.

Issue-Id: SCM-3529
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ie6d5ac9f29a421bb6a9707aebfd0dcb36d416a07