doc: Notify user of FVP system simulation performance impact
As the fvp-base machine is a complete system simulation, users may experience a
performance impact when running the EWAOL image. This patch notifies users of
this in the documentation.

This patch also clarifies references to the supported platforms FVP-Base and
N1SDP, and their associated Yocto MACHINE names fvp-base and n1sdp.

Issue-Id: SCM-3529
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ia1a3435ed1189f9f848c79870328adc86fbea5a6