ewaol-tests: Fix environment variables for BATS logging and cleanup
The shared run-ptest should set the correct log-directory environment variable
appropriate for the package that is being tested. To do this, this patch makes
the test recipes set the environment variable during the package install.

In addition, this patch fixes an error where the BATS scripts were evaluating
the environment clean-up variable incorrectly.

Issue-Id: SCM-3553
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I98e51725dc0d2e841dabfb7a2a26fd5d1b1c613e