ewaol-tests/k3s: Correctly start the k3s service on test suite execution
In order to run the k3s integration tests, the service must be started. This
patch moves the service-start function outside of the optional clean-up
activities, to always run if necessary.

This patch also ensures the service activation is complete before the test-suite
set-up activities attempt to interact with it.

Issue-Id: SCM-3553
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I875e3dc97c5f6c6b372263eb69f7f73eabe5cd4f