tools/qa-checks: If dependency-installation fails, record the check as a FAIL
If a dependency-installation fails, rather than silently dropping the check
module and continuing to evaluate the other checks, the script should record the
check as having failed, outputting the failure in the log appropriately and
returning a non-zero error code.

As the dependency-installation is performed by the virtual environment class
external to the final script execution, the script is made aware of the failure
via the new --skip command line parameter which provides the name of skipped
(due to set-up failure) modules.

A new log formatter is provided to correctly record failure from the context of
a target check module, even if it was not executed.

Issue-Id: SCM-3553
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ia8e94fbb061536886df819866c3d0a2ddd38785b