tools/qa-checks: Add layer-check module, and exclude it from default checks
This QA-check module runs the Yocto Project's checker script 'yocto-check-layer'
on the Yocto layers provided by the repository, in order to validate their
compatibility with the Yocto Project. The check uses the kas-runner.py tool to
execute the yocto-check-layer script within a kas Docker container.

As it takes several minutes to run, the QA check is not included in the checks
by default.

Issue-Id: SCM-3559
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ib9987288817ee6e6ba5c7f72dea89b9e898d57b0