ewaol-tests: Add virtualization runtime integration tests
This patch adds a new BATS- and Expect-based test suite that provides
integration tests for the ewaol-virtualization DISTRO_FEATURE which implements
virtualization in EWAOL via the Xen hypervisor.

The integration test suite runs on the Host via the ptest framework, and
validates that a virtualized Guest is running and has network accessibility, as
well as that the Host image is able to manage the Guest successfully.

Issue-Id: SCM-3583
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I935f5a31e178739ff8aad2f285bb4e3f9c313e72