tools/qa-checks: Improve the layer check's error logging
The layer check expects to read well-formatted passes or failures from
yocto-layer-check. However, if there is an unexpected error then the check
should be verbose, even in non-debug mode.

This patch ensures the output includes the precise command that was executed,
as well as fixes a bug where the output of the subprocess was not being
logged correctly in non-debug mode.

Issue-Id: SCM-3553
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ibc81dea1fc9dfc0e5da48d4e81dc5bb333450967