ewaol-tests: Refactor k3s integration tests
This patch refactors the k3s integration test suite in order to reduce code
duplication by utilizing common code provided by the container engine refactor
patch, as well as refactoring all timeout-related k3s tests to use a single
function which implements a conditional timeout via a callback. A timeout is
also added to the image version check in order to allow sufficient time for the
orchestrator to complete the upgrade.

In addition, complex conditional tests are moved out of the BATS script and into
particular k3s-specific functions that can only return a pass or fail.

Issue-Id: SCM-3583
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I943934edfcc0a771b260d52df6edf059d8b4d999