ewaol-tests: Adapt k3s integration tests for ewaol-virtualization images
This patch enables the k3s integration test suite to run on a virtualization
enabled EWAOL image. To do this, when the ewaol-virtualization DISTRO_FEATURE is
enabled then:

* Relevant k3s test functions are overriden via a bash file that provides
  virtualization-specific function implementations.
* The deployed workload is adapted to constrain pod scheduling to the VM only,
  in order to test deployment of pods on the VM and not the Host.

As both the k3s and Xen virtualization integration test suites require various
different functions to be executed on the guest, the Expect script is adapted to
allow flexibility in the provided command. In addition, short pauses are
added after sending commands from the Expect scripts, to allow the VM some
buffer time to complete the send before performing a subsequent expect.

Due to the k3s test workload being deployed on a 'remote' VM when testing a
virtualization-enabled image, this patch changes how the deployment is accessed
within the test suite. The k3s server now handles requests to the test
deployment (wherever it might be scheduled witin the cluster), via
server-managed routing rules set up using the port-forward command. This command
is run as a daemon via a temporary systemd service on the server node, which is
created during the test suite execution.

When running the k3s test suite on a virtualization-enabled image, test suite
setup code applies a systemd override to the k3s agent on the VM which provides
the necessary details for it to connect to the server on the Host. Teardown
code (if enabled) will then disconnect the agent and return the VM to its
original idle state.

Issue-Id: SCM-3583
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I5658a7633e689a73fb42ecfbf21c7b7eb9f05fb4