ewaol-config: Allow user to pass additional env variables to kas
This patch updates config files for images including ewaol-virtualization
DISTRO_FEATURE. Now it is possible to customize memory and storage amount for
host and VM images. In addition for the VM user can set the amount of available
cpus. Here is a list of variables user can customize:
 (*) EWAOL_VM_MEMORY_SIZE
 (*) EWAOL_VM_NUMBER_OF_CPUS
 (*) EWAOL_VM_ROOTFS_EXTRA_SPACE
 (*) EWAOL_HOST_ROOTFS_EXTRA_SPACE
 (*) EWAOL_HOST_MEMORY_SIZE

This patch also extends kas-runner CI configuration file with two builds:
 (*) virtualization.yml:tests.yml:ci.yml
 (*) sdk.yml:virtualization.yml:ci.yml

Issue-Id: SCM-3722
Signed-off-by: Kamil Dziezyk <kamil.dziezyk@arm.com>
Change-Id: I4dd314eed192d50edbd05284ce45c4484773d3c0