tools/qa-checks: Consider latest git commit for matching file's copyright year
When determining the validity of the copyright year as stated in a file's
header, the header check now considers the date of the latest git commit that
modified that file. If the file has local changes that have yet to be
committed, or is not yet tracked by git at all, then the check continues to use
the file's latest modification date as given by the local filesystem.

Issue-Id: SCM-3910
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ic8ec4d7ede0f47eb22f3852d249e452a068f5ee6