tools: Update required versions of system dependencies
This patch updates the required version of Python for running the tools from 3.6
to 3.8, and also defines the required version of Git for running the licence and
copyright header check, as 2.25. These are defined in both the documentation and
as part of initial run-time validation steps in the tooling code.

Issue-Id: SCM-3910
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ibccf10e9f9055c43424eb92d39504b9e9fcb8a13