ewaol-distro: Update config search path for kernelcfg_check
This patch extends the search mechanism for the required kernel configuration
file.

If the required config fragment is present under the 'yocto-kernel-cache' repo,
then other configs from all included layers are ignored. If the required config
is not present in the above repo, 'STAGING_KERNEL_DIR' is being searched for
the required file.

This patch also fixes a bug in config ignore mechanism and changes the md5sum
of 'docker.cfg'.

Issue-Id: SCM-3910
Signed-off-by: Kamil Dziezyk <kamil.dziezyk@arm.com>
Change-Id: I2269d4bc96e6c33b3e411feb0b8132a81ad49969