tools/qa-checks: Fix bug getting the last modification date of external works
A previous patch added new functionality to the copyright date checking, to
consider the date of the latest git commit that modified the file. This should
have applied both to internal works and modified external works, but was only
applied to the former. This patch ensures it applies to both.

Issue-Id: SCM-3910
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I5bce036c154a7a0e34817c48a8884a5527526810