ewaol-tests,doc: Run container engine tests on VM as part of the Host test suite
The container engine integration tests should be run on both the Host and the
VM. Rather than the user having to do this manually, this patch includes an
extra test on the Host that logs into the VM and executes the test suite. This
extra test will fail if the test suite fails on the VM.

This patch also increases the Expect script's timeout, as running the test suite
on the VM takes longer than any previously executed command.

As the test suite is identical on both the Host and the VM, the test includes a
conditional to determine if it is running on the Host or the VM. If running on
the VM, the test is simply skipped.

Issue-Id: SCM-3910
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ia719f2a72197a354ecc37b583802c49f19c7c04d