ewaol-tests: Fix Expect log-in procedure, and remove home directory constraint
When the Expect script logs into the VM for the first time, it should wait for
the boot process to complete before running any expectations/commands. This
patch implements this by waiting for all output to stop (with a given timeout)
before proceeding, following the console spawn.

In addition, if the user has logged into the VM and changed the working
directory, this shouldn't stop the expect script from subsequently working.
Further, changing the working directory should be an allowable operation for the
expect script to perform.

Issue-Id: SCM-3910
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ic37755430a930d787d843b9bbfae2c767def0a10