tools/qa-checks: Make copyright header check use GIT_AUTHOR_DATE
A patch's GIT_COMMITTER_DATE updates when it is (for example) cherry-picked,
causing the copyright header check to potentially prompt for an updated
copyright header. As copyright applies to the authored content of a patch, this
patch makes the header check consider its GIT_AUTHOR_DATE instead, which reports
the date that the content was authored.

The ewaol-vm-image.bb recipe is an edge case that was authored in 2021 but made
to include copyright for 2022 (which was when it was merged), due to the header
check's usage of GIT_COMMITTER_DATE. Now that the check is using
GIT_AUTHOR_DATE instead, this patch makes a trivial change to that file so that
the latest author date of the recipe is reported as 2022, fixing the error
produced by the header check.

Issue-Id: SCM-3910
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I9c072e7ca668d749ded2ad1797b6cf5a1cf3499c