ewaol-[distro,config],doc: Refactor EWAOL build targets and kas config files
This patch refactors the naming scheme for EWAOL build targets, and creates the
EWAOL Virtualization SDK image. The build targets are now as follows:

Baremetal Images:
  * ewaol-baremetal-image
  * ewaol-baremetal-sdk-image

Virtualization Images:
  * ewaol-virtualization-image
  * ewaol-virtualization-sdk-image

In addition, this patch refactors the kas config files to more clearly define
the process for building the different targets via kas. It is now necessary to
supply kas with an architecture config file (e.g. baremetal.yml) and a machine
config file (e.g. n1sdp.yml) for a valid build. To support this, the
`ewaol-baremetal` DISTRO_FEATURE is introduced as a counterpart to the
`ewaol-virtualization` DISTRO_FEATURE.

Issue-Id: SCM-3839
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I66b27bd55c8762cc3a389e0150554c536855bab9