tools/kas-runner: Fix file path resolution
The common directories containing all kas config files and the
project root to run is added as a read only volume for the
container. Kas file paths are modified to point to the mount path
to pass to the container.

This allows configs outside project root to be used without extra
manual configuration and absolute paths to be used. Relative paths
will still be considered relative to project root.

Issue-Id: SCM-3900
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I1a5aeebdb9b12e5ca4e3f1cdc6434fbaf4872c15