tools/qa-checks,ewaol-config: Fix spell check parsing, regenerate dictionary
Include periods (.) and apostrophies (') when parsing words for the spell check,
so that "e.g." gets parsed instead of "e" and "g".

This patch also fixes an issue with words parsed from the custom dictionary,
where hyphens (for example) were used as a delimiter instead of being part of
the custom word.

As the check is case-insensitive, the output from the spell check is now lower
case and all words are stored in lower case in the custom dictionary, in order
to make this clearer.

Following the documentation refactor and spell check changes, the custom
dictionary has been re-generated from the current documentation.

Issue-Id: SCM-4394
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I8f9607257a4903ef4c125e235d220a16a695584e