ewaol-tests: Store test pid files in volatile path /var/run
To avoid multiple concurrent runs of a test suite the runtime
integration tests store temporary PID files to disk during
execution. These are normally deleted after a successful run.
However, if the system is rebooted (for example after a system
crash) then the files persist but the running test does not.

To prevent the files persisting after a reboot the pid files are
now stored in volatile memory (/var/run).

Issue-Id: SCM-4420
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: Ib5d5330ed6ea09d3418d2343362b839a7c5d71bb