doc: Update documentation for contributing and submitting issues
This patch provides more precision about the minimal requirements for
contributions, and clarifies that the automated QA checks can only validate a
subset of those requirements.

In addition, the suggested reporting of Gitlab issues is expanded from relating
to bug-reports only, to also include security concerns and feature requests.

The contributing page is also slightly restructured for more clarity.

Issue-Id: SCM-4394
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I40b0d1daff14e1db0ff2be4aabe5bf2794c379cb