tools/qa-checks: Support optional variables for the QA-checks
With this patch, a check module can declare optional variables. If these
variables are not defined (either via the command-line or via a config file),
then the relevant check module will be invoked with the varible set as None.

If a variable is not declared as optional then the previous behavior remains,
where it is considered required. If any required variable is not defined then
the check will be dropped, and a warning output to the user.

Issue-Id: SCM-3902
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I7617ad0e5f36c0ca068aee855ed00368207aa30c