tools/qa-checks,doc: Add a non-inclusive language check module
This check aims to identify opportunities across the target repository to use
more inclusive language. Potentially non-inclusive language is reported as a
result of the tool, along with the filename and line number so that
consideration can be made for alternatives. Exceptions can be provided when
appropriate alternatives cannot be found, by tagging it with
"inclusivity-exception" before the usage.

The list of non-inclusive language is stored within the project configuration
files.

Issue-Id: SCM-4425
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Id1f2ba92894274cdde2d1464c62456aba060c4dd