ewaol-tests: Use BATS 1.5.0 version
With v1.6.0, BATS returns success with failing tests. The exit status of a
BATS 'run' command called from the teardown_file() function is returned as
the exit status of the bats script. This means that even if the test cases
fail, a successful teardown will produce a passing test suite execution.
This is a regression from v1.5.0.
https://github.com/bats-core/bats-core/issues/575

Issue-Id: SCM-4394
Signed-off-by: Diego Sueiro <diego.sueiro@arm.com>
Change-Id: I7501adf4a65e0271bc54c8693a794c5dd9c31ce6