doc/variables: Refactor definition of link variables
Rather than forcing links and simple string substituions to be partitioned into
two separate Python dictionaries (even if the two dictionaries apply to the same
documentation component), make organizing the variables easier by allowing a
single dictionary to include both links and simple string substitutions.

To do this, a variable key is prefixed with "link:" to process it as a link.

In addition, as documentation for old releases (migration guidance, changelogs,
etc) may require variable definitions which point to older versions of internal
and external links, this patch provides a new dict where per-release definitions
can be provided.

Issue-Id: SCM-4394
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I288b87b99e1e6be7f50380808157db8ff6aff5e3