ewaol-tests: If a test suite fails, print the log-file to stdout for debugging
To do this, print the log from the generic run-* script that executes each test
suite. As this generic script doesn't know what test suite it will be running,
or what the user might customize that test suite's log file to be, use a
hard-coded temporary log-file placed in the same volatile runtime directory that
is used for storing the test PIDs, and log messages to that file simultaneously
with the normal logging procedure.

Also, this patch allows setting extra parameters to be passed to bats "run-.*"
scripts via the 'EWAOL_TEST_OPTIONS' build-time variable. Currently, it is set
to "--show-output-of-passing-tests" to give the user more information about
what the test suite is doing at run-time.

Furthermore, spaces are replaced with tabs in the 'run-test-suite' file to allow
better code readability with here-document indentions, along with two more
variables: 'TEST_RUNTIME_DIR' and 'EWAOL_TEST_BATS_OPTIONS', are set with the
envsubst in the test recipes.

Issue-Id: SCM-4394
Signed-off-by: Richard Neill <richard.neill@arm.com>
Signed-off-by: Kamil Dziezyk <kamil.dziezyk@arm.com>
Change-Id: I384d2f3046d8195b0bb26ce61d175ddd46323c3f