ewaol-tests,doc: Fix race condition in K3s integration tests logic
When the test deployment is upgraded to a later container image version, K3s
stops the Pod(s) with the old container image and starts the Pod(s) with the new
container image. The step after upgrading the deployment is to query the running
Pods and check if they are using the new container image.

This situation represents a race condition. If the query is performed quickly,
then the old Pod hasn't yet been stopped and the old Pod name is considered to
be the new Pod. This causes an error later in the tests when it does stop, and
subsequent queries to it become invalid.

This patch fixes the race condition by ensuring that a full set of 'new' Pods
are running after the upgrade, before continuing with the rest of the test
suite.

The documentaton is updated to reflect the new logic.

Issue-Id: SCM-4394
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I8d9020097897ed5cd90cf6cee0c808ce6c3502b1