ewaol-bsp/systemd: Drop i40e network device systemd override on Guest VM
The systemd override for systemd-networkd-wait-online causes all i40e based PCI
network adapters to be ignored. If this is the only type of network device, then
the systemd-networkd-wait-online service will time-out, and any dependent
services will fail to start.

As this can happen on EWAOL Guest VMs which don't configure a Xen bridge network
device and instead use a i40e PCI network adapter via PCI passthrough, this
patch ensures that the systemd override is not installed on Guest VMs.

Issue-Id: SCM-4754
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I80541673de1fee5b07f9846fade9a48b9c63e9ca