tools/kas-runner: Fix artifacts deployment with circular symlinks
Recipes which produce directories in the build folder which are symlinks to
their own parent directories cause the artifacts deployment to fail when the
build directories are parsed. This is because the glob implementation from the
'glob' package enters an infinite loop when operating in recursive mode.

Instead use the glob implementation from the 'pathlib' package which is already
imported, as this is capable of handling circular symlinks.

Also, during debugging of this infinite loop it was noticed that it was unclear
what the tool was working on in the terminal but with the log_file option
enabled. Therefore, tweak the logging to output more informative messages both
in the log file and in the terminal.

Issue-Id: SCM-4754
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Id932c6a2e653353ca1c622c1b53842e6b62b58e5