tools/kas-runner: Support interactive kas commands via Python pseudo-terminal
A user may want to run the kas command interactively, for example when launching
kas shell to enter the build environment. This patch provides proper support for
this functionality via the Python pseudo-terminal (pty) package. If the command
is not run interactively, the standard subprocess-based implementation is used.

As the pty spawn call must be passed an executable command directly as the first
argument, environment variable definitions cannot be prepended, where instead
the command inherits the environment of the caller. Therefore define environment
variables in advance via os.environ, when running the tool in no_containerize
mode

Issue-Id: SCM-4754
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ia89a59f3070ed8f2f2d26d69bc89d66ca7a723fe