tools/kas-runner: When running kas shell, make container interactive by default
For convenience, if the user supplies "shell" as the value for kas_arguments
and it is a containerized execution, simply append "-it" to the engine_arguments
value to avoid the user having to do it manually.

As commands such as "shell -c 'bitbake-getvar ABC'" are valid but don't require
interactivity, only conservatively append the interactive container options if
"shell" is set on its own.

Issue-Id: SCM-4754
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: I7c0e143e3343ffe3604425673bb35e0647c05d99