ewaol-config/qa-checks: Ensure original works within project use MIT license
A previous patch changed the license and copyright header check to accept all
SPDX license identifiers as valid licenses for project files by default.

Follow the validation that was in place before that patch by ensuring that
original works contributed to the project are licensed with the MIT license,
re-aligning the validation with the current contribution guidelines.

Issue-Id: SCM-5249
Signed-off-by: Richard Neill <richard.neill@arm.com>
Change-Id: Ibd1091092e2f294a3c7375ee20fc104b8b0846f6